requirements.txt
langcahin
langchain_community
python-dotenv
openai